usb: dwc2: gadget: stop isoc ep in transfer when disable ep
When disable ep, force to stop isoc ep in transfer even if
the ep has been disabled. Without this patch, it's easily
to make the dwc2 controller enter an abnormal state, and
both the ep0 and the ep in can't work when enable the eps
again next time.

Change-Id: Id9b31bd7a93e138261b9c2a6c3e5e42fb3bde83e
Signed-off-by: William Wu <william.wu@rock-chips.com>

rockchip 0ba984903508463186626dcb82a8f6192665eef1